Change check to this.nestDepth > 0 instead of 1
Whenever depth is set as all, we run our code, which appends additional columns to the data (KAFKA_OFFSET, KAFKA_TIMESTAMP, etc...). This allows our branch to run and run the code that adds those columns to the columnAppend code, otherwise it fails to infer the type of KAFKA_OFFSET and KAFKA_TIMESTAMP and the run fails entirely.