Perform cleanup
We no longer need to check this.nestDepth as the logic is the same for both the case where further flattening is needed, and otherwise. In either case we need to check if their are any additional columns in the message compared to the snowflake table.

The check before was a bandaid due to the fact the schematization wasn't working properly.